+++ react/components/CodeConfirmation.js
@@ -1,13 +1,16 @@
 import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
-import { Input, Button, IconArrowBack } from 'vtex.styleguide'
 import { injectIntl, intlShape } from 'react-intl'
+
+import { Input, Button } from 'vtex.styleguide'
+import { AuthState, AuthService } from 'vtex.auth'
 
 import { translate } from '../utils/translate'
 import { isValidAccessCode } from '../utils/format-check'
-import { AuthState, AuthService } from 'vtex.auth'
 import Form from './Form'
 import FormError from './FormError'
+import GoBackButton from './GoBackButton'
+
 
 /** CodeConfirmation tab component. Receive the code from an input and call the signIn API */
 class CodeConfirmation extends Component {
@@ -76,20 +79,7 @@ class CodeConfirmation extends Component {
         }
         footer={
           <Fragment>
-            <div className="vtex-login__back-button">
-              <Button
-                variation="tertiary"
-                size="small"
-                onClick={() => onStateChange({ step: previous })}
-              >
-                <span className="vtex-login__back-icon c-link">
-                  <IconArrowBack size={10} color="currentColor" />
-                </span>
-                <span className="t-small ml2">
-                  {translate('login.goBack', intl)}
-                </span>
-              </Button>
-            </div>
+            <GoBackButton onStateChange={() => onStateChange({ step: previous })} />
             <div className="vtex-login__send-button">
               <AuthService.LoginWithAccessKey
                 onSuccess={this.handleSuccess}