+++ react/components/LoginComponent.js
@@ -1,15 +1,17 @@
 import React, { Component, Fragment } from 'react'
 import { Link } from 'render'
-import { Button } from 'vtex.styleguide'
 import OutsideClickHandler from 'react-outside-click-handler'
 
-import LoginContent from '../LoginContent'
+import { ButtonWithIcon } from 'vtex.styleguide'
 import Icon from 'vtex.use-svg/Icon'
+
+import LoginContent from '../LoginContent'
 import { truncateString } from '../utils/format-string'
 import { translate } from '../utils/translate'
 import { LoginPropTypes } from '../propTypes'
 import { getProfile } from '../utils/profile'
 
+const profileIcon = iconSize => (<Icon id="hpa-profile" size={iconSize} />)
 export default class LoginComponent extends Component {
   static propTypes = LoginPropTypes
 
@@ -25,7 +27,6 @@ export default class LoginComponent extends Component {
     const {
       iconSize,
       iconLabel,
-      iconClasses,
       labelClasses,
       intl,
       renderIconAsLink,
@@ -37,9 +38,6 @@ export default class LoginComponent extends Component {
 
     const iconContent = (
       <Fragment>
-        <div className={`${iconClasses}`}>
-          <Icon id="hpa-profile" size={iconSize} />
-        </div>
         {profile ? (
           <span className={`vtex-login__profile order-1 f6 pl4 ${labelClasses} dn-m db-l`}>
             {translate('login.hello', intl)}{' '}
@@ -64,7 +62,7 @@ export default class LoginComponent extends Component {
     }
 
     return (
-      <Button variation="tertiary" icon onClick={onProfileIconClick}>
+      <ButtonWithIcon variation="tertiary" icon={profileIcon(iconSize)} iconPosition="left" onClick={onProfileIconClick}>
         <div
           className="flex pv2 items-center"
           ref={e => {
@@ -73,7 +71,7 @@ export default class LoginComponent extends Component {
         >
           {iconContent}
         </div>
-      </Button>
+      </ButtonWithIcon>
     )
   }