+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Use `ButtonWithIcon` in favour of `Button` from styleguide. 
+
+### Fixed
+- Create `GoBackButton` to remove duplicated code.
 
 ## [2.2.0] - 2018-12-17
 ### Added